Fully transitioned to SASS for styling
I have many colors based on just one color, for hovered elements/active tabs, that now update automatically based on a single color variable thanks to sass. This eliminates the need for manual changes to each color when I inevitably change the theme.

Nesting is also nice, and everything is more clear to me as a developer-user